[breaking] Allow specifying the kafka partition
This allows using different partition keys but still send the records to
the same partition. E.g. I want all the records from the same account to
be in the same partition, but I want to use the user id as the key.

This also helps to make the goharvest more effecient because now we can
have more distributed load even though the records still end up in the
same partitions.

**BREAKING CHANGE**: `kafka_partition` field must be added to the outbox
table before using this code.